refactor: use listeners
