--- services/bot/src/lib/internal/structures/modules/ModuleStore.ts
@@ -1,8 +1,0 @@
-import { Store } from "@sapphire/framework";
-import { Module } from "./Module";
-
-export class ModuleStore extends Store<Module> {
-	public constructor() {
-		super(Module, { name: "modules" });
-	}
-}
